# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Aurora Admin API",
        "title": "Aurora Admin",
        "contact": {},
        "version": "1.0"
    },
    "host": "0.0.0.0:5800",
    "basePath": "/",
    "paths": {
...
        "/api/{service}/status": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "service_status"
                ],
                "summary": "获取状态",
                "parameters": [
                    {
                        "enum": [
                            "dontstarve"
                        ],
                        "type": "string",
                        "description": "service",
                        "name": "service",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: Inactive, Active, Starting, Stopping, WaitingActive, Abnormal",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/comety.Status"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "comety.Status": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}